+++ src/settings.js
@@ -1,9 +1,19 @@
 import { MODULE_NAME } from "./const.js";
 
 export function registerSettings() {
+
+    game.settings.register(MODULE_NAME, "showButton", {
+        name: "AIP.showButton.name",
+        hint: "AIP.showButton.hint",
+        scope: "client",
+        config: true,
+        type: Boolean,
+        default: true,
+    });
+
     game.settings.register(MODULE_NAME, "debug", {
-        name: "AIP.DebugName",
-        hint: "AIP.DebugHint",
+        name: "AIP.debug.name",
+        hint: "AIP.debug.hint",
         scope: "client",
         config: true,
         type: Boolean,